Reduce sizebuf_t size.

+++ src/client/parse.c
@@ -312,7 +312,7 @@ static void CL_ParseFrame(int extrabits)
         }
     }
 
-    SHOWNET(2, "%3zu:playerinfo\n", msg_read.readcount - 1);
+    SHOWNET(2, "%3u:playerinfo\n", msg_read.readcount - 1);
 
     // parse playerstate
     bits = MSG_ReadWord();
@@ -361,7 +361,7 @@ static void CL_ParseFrame(int extrabits)
         }
     }
 
-    SHOWNET(2, "%3zu:packetentities\n", msg_read.readcount - 1);
+    SHOWNET(2, "%3u:packetentities\n", msg_read.readcount - 1);
 
     CL_ParsePacketEntities(oldframe, &frame);
 
@@ -372,7 +372,7 @@ static void CL_ParseFrame(int extrabits)
     if (cl_shownet->integer > 2) {
         int seq = cls.netchan.incoming_acknowledged & CMD_MASK;
         int rtt = cls.demo.playback ? 0 : cls.realtime - cl.history[seq].sent;
-        Com_LPrintf(PRINT_DEVELOPER, "%3zu:frame:%d  delta:%d  rtt:%d\n",
+        Com_LPrintf(PRINT_DEVELOPER, "%3u:frame:%d  delta:%d  rtt:%d\n",
                     msg_read.readcount - 1, frame.number, frame.delta, rtt);
     }
 #endif
@@ -1219,12 +1219,12 @@ CL_ParseServerMessage
 void CL_ParseServerMessage(void)
 {
     int         cmd, index, extrabits;
-    size_t      readcount;
+    uint32_t    readcount;
     uint64_t    bits;
 
 #if USE_DEBUG
     if (cl_shownet->integer == 1) {
-        Com_LPrintf(PRINT_DEVELOPER, "%zu ", msg_read.cursize);
+        Com_LPrintf(PRINT_DEVELOPER, "%u ", msg_read.cursize);
     } else if (cl_shownet->integer > 1) {
         Com_LPrintf(PRINT_DEVELOPER, "------------------\n");
     }
@@ -1238,7 +1238,7 @@ void CL_ParseServerMessage(void)
     while (1) {
         readcount = msg_read.readcount;
         if (readcount == msg_read.cursize) {
-            SHOWNET(1, "%3zu:END OF MESSAGE\n", readcount);
+            SHOWNET(1, "%3u:END OF MESSAGE\n", readcount);
             break;
         }
 
@@ -1249,7 +1249,7 @@ void CL_ParseServerMessage(void)
         extrabits = cmd >> SVCMD_BITS;
         cmd &= SVCMD_MASK;
 
-        SHOWNET(1, "%3zu:%s\n", msg_read.readcount - 1, MSG_ServerCommandString(cmd));
+        SHOWNET(1, "%3u:%s\n", msg_read.readcount - 1, MSG_ServerCommandString(cmd));
 
         // other commands
         switch (cmd) {
@@ -1364,7 +1364,7 @@ void CL_ParseServerMessage(void)
 
         // if recording demos, copy off protocol invariant stuff
         if (cls.demo.recording && !cls.demo.paused) {
-            size_t len = msg_read.readcount - readcount;
+            uint32_t len = msg_read.readcount - readcount;
 
             // it is very easy to overflow standard 1390 bytes
             // demo frame with modern servers... attempt to preserve
@@ -1398,7 +1398,7 @@ bool CL_SeekDemoMessage(void)
 
 #if USE_DEBUG
     if (cl_shownet->integer == 1) {
-        Com_LPrintf(PRINT_DEVELOPER, "%zu ", msg_read.cursize);
+        Com_LPrintf(PRINT_DEVELOPER, "%u ", msg_read.cursize);
     } else if (cl_shownet->integer > 1) {
         Com_LPrintf(PRINT_DEVELOPER, "------------------\n");
     }
@@ -1411,12 +1411,12 @@ bool CL_SeekDemoMessage(void)
 //
     while (1) {
         if (msg_read.readcount == msg_read.cursize) {
-            SHOWNET(1, "%3zu:END OF MESSAGE\n", msg_read.readcount);
+            SHOWNET(1, "%3u:END OF MESSAGE\n", msg_read.readcount);
             break;
         }
 
         cmd = MSG_ReadByte();
-        SHOWNET(1, "%3zu:%s\n", msg_read.readcount - 1, MSG_ServerCommandString(cmd));
+        SHOWNET(1, "%3u:%s\n", msg_read.readcount - 1, MSG_ServerCommandString(cmd));
 
         // other commands
         switch (cmd) {
